mcp_server/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
snowflake-connector-python==3.11.0
requests==2.32.3
